Added Label Shrines, trouble with Music
1. Added a Label to flash all the shrines I have collected.  We can
discuss this on Friday, remove objectionable content, etc. I thought it
might be nice for our dying cultist main character to have his life
flash before his eyes before moving the the mythical Virus World.
2. Moved music referenced in the game to the Game folder.  I had trouble
getting it to reference the location of the subfolder.  I'll try to test
some more this week but if anyone else can figure it out, that would be
cool.  I wouldn't sweat it for now.