Proposing a change of internal names
I prefer to use human-readable names rather than two-letter
abbreviations. It's just much easier to read, and easier to understand.
The old names still work, so it won't break the code.
Feel free to revert this if you want to keep the old names, but I feel
like this will make the code more readable.